Fix installation via pip when pypandoc isn't available
The README.md file is converted to reStructuredText when uploading the
package to PyPI, because they don't support markdown. The conversion
uses pypandoc and use to fail when pypandoc wasn't installed.

Make this conditional on the availability of the pypandoc package.
Requiring the installation of pypandoc (quite large) just for that
purpose would be too much.