Build the list method on Linux only.
There is no portable way to list SHM segments, that's more or less a
flaw in the SHM API. On Linux SHM segments are implemented as regular
files on a tmpfs filesystem, usually mounted as /dev/shm. The list
method scans this mount point and looks for SharedArray headers. This
won't work on other operating systems (*BSD, Darwin, Solaris, etc.),
so disable the method on non-Linux targets.