Use mmap only, no reads or seek
At least on a mac `lseek` and `read` fail when operating on a shared memory file
descriptor. Seems like POSIX only promises resize and memory map, but not read
or seek.

Also on a mac resizing shared memory file rounds up to nearest 4K, so we have to
query actual file size after resize before writing header information to the end
of the file.